unbroke the best minimum confirmed interview.

+++ src/main/java/io/github/ryan_glgr/hansel_grapher/TheHardStuff/Interview.java
@@ -248,7 +248,7 @@ public class Interview {
     }
 
     private static void binarySearchStringOfExpansionsInterview(ArrayList<Node> allNodes, Node rootNode) {
-        int totalNodes = (int) allNodes.size();
+        int totalNodes = allNodes.size();
         int questionsAsked = 0;
 
         while (true) {
